+++ backend/elasticsearch/src/main/java/org/hibernate/search/backend/elasticsearch/search/aggregation/impl/ElasticsearchSearchAggregationBuilderFactory.java
@@ -114,7 +114,7 @@ public class ElasticsearchSearchAggregationBuilderFactory
 
 		@Override
 		public ElasticsearchFieldAggregationBuilderFactory extractComponent(ElasticsearchIndexSchemaFieldNode<?> schemaNode) {
-			return schemaNode.getAggregationBuilderFactory();
+			return schemaNode.getType().getAggregationBuilderFactory();
 		}
 
 		@Override
