+++ backend/elasticsearch/src/main/java/org/hibernate/search/backend/elasticsearch/search/predicate/impl/ElasticsearchSearchPredicateBuilderFactoryImpl.java
@@ -198,7 +198,7 @@ public class ElasticsearchSearchPredicateBuilderFactoryImpl implements Elasticse
 
 		@Override
 		public ElasticsearchFieldPredicateBuilderFactory extractComponent(ElasticsearchIndexSchemaFieldNode<?> schemaNode) {
-			return schemaNode.getPredicateBuilderFactory();
+			return schemaNode.getType().getPredicateBuilderFactory();
 		}
 
 		@Override
