+++ backend/elasticsearch/src/main/java/org/hibernate/search/backend/elasticsearch/search/projection/impl/ElasticsearchSearchProjectionBuilderFactory.java
@@ -182,7 +182,7 @@ public class ElasticsearchSearchProjectionBuilderFactory implements SearchProjec
 
 		@Override
 		public ElasticsearchFieldProjectionBuilderFactory extractComponent(ElasticsearchIndexSchemaFieldNode<?> schemaNode) {
-			return schemaNode.getProjectionBuilderFactory();
+			return schemaNode.getType().getProjectionBuilderFactory();
 		}
 
 		@Override
